GitHub action to verify WCP manifests
This patch adds support for a GitHub action that verifies the WCP
manifests by deploying a Kind cluster and ensuring the WCP manifests
can be applied to the cluster successfully.

+++ hack/deploy-local.sh
@@ -7,20 +7,30 @@
 set -o errexit
 set -o pipefail
 set -o nounset
+set -x
 
-YAML=$1
+# VERIFY_MANIFESTS is set to a non-empty value if the only thing this file
+# should be doing is verifying the YAML can be successfully applied to a Kind
+# cluster. When this is set, we do not wait to verify that deployments,
+# webhooks, pods, etc. are online. The only thing about which we care is if
+# "kubectl apply" succeeded.
+VERIFY_MANIFESTS="${VERIFY_MANIFESTS:-}"
+
+YAML="${1}"
 KUBECTL="kubectl"
 
 VMOP_NAMESPACE="vmware-system-vmop"
 VMOP_DEPLOYMENT="vmware-system-vmop-controller-manager"
 
 DEPLOYMENT_EXISTS=""
-if $KUBECTL get deployment -n ${VMOP_NAMESPACE} ${VMOP_DEPLOYMENT} >/dev/null 2>&1; then
-  DEPLOYMENT_EXISTS=1
+if [ -z "${VERIFY_MANIFESTS:-}" ]; then
+  if $KUBECTL get deployment -n ${VMOP_NAMESPACE} ${VMOP_DEPLOYMENT} >/dev/null 2>&1; then
+    DEPLOYMENT_EXISTS=1
+  fi
 fi
 
-# Deploy and check cert-manager
-CERTMANAGER_NAMESPACE="vmware-system-cert-manager"
+# Check to see if cert-manager exists.
+CERTMANAGER_NAMESPACE="${CERTMANAGER_NAMESPACE:-cert-manager}"
 CERTMANAGER_DEPLOYMENTS=(
   cert-manager
   cert-manager-cainjector
@@ -33,25 +43,34 @@ if $KUBECTL get deployment -n "${CERTMANAGER_NAMESPACE}" "${CERTMANAGER_DEPLOYME
 fi
 if [[ -z $CERTMAN_EXISTS ]]; then
   ./hack/deploy-local-certmanager.sh
+
   for dep in "${CERTMANAGER_DEPLOYMENTS[@]}"; do
     $KUBECTL rollout status -n "${CERTMANAGER_NAMESPACE}" deployment "${dep}"
   done
 
   # TODO Find a better way to wait for this...
-  echo $'\nSleeping for 60s - waiting for webhooks to be initialized\n'
+  echo $'\nSleeping for 1m - waiting for cert-manager webhooks to be initialized\n'
   sleep 60
 fi
 
-# Hack to reduce the number of replicas deployed from 3 to 1
-# when deploying onto a single node kind cluster.
-NODE_COUNT=$(kubectl get node --no-headers 2>/dev/null | wc -l)
-if [ "$NODE_COUNT" -eq 1 ]; then
-  sed -i -e 's/replicas: 3/replicas: 1/g' "$YAML"
-  # remove the generated '-e' file on Mac
-  rm -f "$YAML-e"
+if [ -z "${VERIFY_MANIFESTS:-}" ]; then
+  # Hack to reduce the number of replicas deployed from 3 to 1
+  # when deploying onto a single node kind cluster.
+  NODE_COUNT=$(kubectl get node --no-headers 2>/dev/null | wc -l)
+  if [ "$NODE_COUNT" -eq 1 ]; then
+    sed -i -e 's/replicas: 3/replicas: 1/g' "$YAML"
+    # remove the generated '-e' file on Mac
+    rm -f "$YAML-e"
+  fi
 fi
 
+# Apply the infrastructure YAML.
 $KUBECTL apply -f "$YAML"
+
+# If VERIFY_MANIFESTS is set then we can go ahead and exit.
+if [ -n "${VERIFY_MANIFESTS:-}" ]; then
+  exit 0
+fi
 
 if [[ -n $DEPLOYMENT_EXISTS ]]; then
   $KUBECTL rollout restart -n ${VMOP_NAMESPACE} deployment ${VMOP_DEPLOYMENT}
@@ -63,7 +82,7 @@ until $KUBECTL wait --for=condition=Ready -n vmware-system-vmop cert/vmware-syst
 done
 
 if [[ $# -eq 2 ]]; then
-  VMCLASSES_YAML=$2
+  VMCLASSES_YAML="${2}"
   # Hack that retries applying the default VM Classes until the
   # validating webhook is available.
   VMOP_VMCLASSES_ATTEMPTS=0
